requirements.txt
aiogram==3.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
yandexcloud==0.242.0
aiohttp==3.0.5